refactor(stack): move init stack to directory
